Remove conda dependency, update docs
Changed the virtual environment to use pythons venv instead of conda.
Added a requirements.txt with version locks to allow the simple creation
of containers and virtual environments.
The requirements.txt is platform-independend and excludes the installation
of jupyter (which is os-dependent).

Also, updated the docs to incorporate said changes together with some
reorganisation of the docs folder to improve readability.
Add the installation of Jupyter to docs.